Switch tempfile to `with` instead of more manual handling
Tempfile supports the convenient `with` syntax to avoid having to manually
remove the directory once we're done, regardless of intervening errors.

The motivation for this is an error with `save_dash_app` seen in dev (but not
locally) that I can't get any information about because the `except` eats the
error and then just prints a message that it fails. Now errors will bubble all
the way up, and the temp directory will still get removed.

testing: I confirmed that `format(TemporaryDirectory)` returns a string, so
where we're using it for shell commands, it should work. I didn't check the
usage of shutil.copy, but I assume it'll work too. I did not actually try
running this.