Remove default value for title in save_static_image
Title is very much not-optional and validation in save_static_image will fail
if the title is less than 10 characters let alone `None`. As such, remove the
default value to avoid giving the impression that it can be omitted.

This necessarily changes the order of the arguments, so this will break any
users calling this positionally. I leave it to reviewers to decide how to
weight that. I would not be surprised if the internal SDK call was the only
actual usage.

testing: none