Support hero image
Changes save_dash_app to take a (single) PNG image to use as the hero image in
the study space.

Support for not including an image and using a default has still not been added.

testing: Passing in a PNG results in a lovely hero image in the study space.