Fix for wrong app.py file in tarball
`find_files` was finding the app.py in the examples directory and including it
in the tarball instead of the app.py passed in by the user. This causes
problems.

Drop usage of find_files and instead pass additional filenames and app filepath
directly to shutil.copy. If any aren't found, shutil will complain.

testing: none